+++ agent/test/CMakeLists.txt
@@ -34,6 +34,7 @@ else()
   set(SRC ${SRC_COMMON}
     check_event_log_data_test.cc
     check_event_log_filter_test.cc
+    check_event_log_uniq_test.cc
     check_windows_cpu_test.cc  
     check_windows_memory_test.cc 
     check_uptime_test.cc
